# internal/httpapi/router.go
package httpapi

import (
	"context"
	"crypto/subtle"
	"errors"
	"log/slog"
	"net"
	"net/http"
	"strings"
	"sync"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/labstack/echo/v4/middleware"

	"github.com/chadhao/jx-procurement-platform/internal/access"
	"github.com/chadhao/jx-procurement-platform/internal/approval"
	"github.com/chadhao/jx-procurement-platform/internal/chain"
	"github.com/chadhao/jx-procurement-platform/internal/config"
	"github.com/chadhao/jx-procurement-platform/internal/flow"
	"github.com/chadhao/jx-procurement-platform/internal/inbox"
	"github.com/chadhao/jx-procurement-platform/internal/objectstore"
	"github.com/chadhao/jx-procurement-platform/internal/observ"
	"github.com/chadhao/jx-procurement-platform/internal/orgsync"
	"github.com/chadhao/jx-procurement-platform/internal/permission"
	"github.com/chadhao/jx-procurement-platform/internal/platform/feishu"
	"github.com/chadhao/jx-procurement-platform/internal/specload"
	"github.com/chadhao/jx-procurement-platform/internal/store"
	fsync "github.com/chadhao/jx-procurement-platform/internal/sync"
	"github.com/chadhao/jx-procurement-platform/internal/worker"
)

// Deps 路由依赖集合（由 cmd 装配）。
type Deps struct {
	Env        *config.Env
	DB         *store.DB
	Log        *slog.Logger
	Metrics    *observ.Metrics
	Health     *observ.Health
	Inbox      *inbox.Service
	Worker     *worker.Worker
	Subscriber *fsync.Subscriber
	// Flow 审批领域服务（架构转向 ③ 上电，T03b/T04b）：页面两键 / 四操作 / 待办 / 入站回调。
	// ★ 回调与页面两键**走同一状态机出口**（docs/11 R11），故共用此一个服务实例。
	Flow *flow.Service
	// ApprovalDefs 三方审批定义注册表（approval.Registry + feishu.ExternalApprovalClient）。
	// 装配即用：定义注册/更新经此实例；生产装载入口＝`POST /api/admin/approval/defs/sync`
	// （docs/16 §2-C 通道①，admin 组 + requireSysAdmin）。
	ApprovalDefs *approval.Registry
	// ApprovalReconciler 审批对账器（T03）：`POST /internal/approval/check` 唯一入口（R24）。
	ApprovalReconciler *fsync.ApprovalReconciler
	Perm               *permission.Loader
	Auth               *access.Authenticator
	Maps               *config.Maps
	WebUI              http.Handler
	Version            string
	// Feishu 飞书客户端：附件**按需拉取**需要（B39）。为 nil 时下载端点返回 502。
	Feishu feishu.Client
	// Contact 身份转换端口（user_id → open_id，docs/16 §2-A-3）：回调官方发 user_id
	//（租户内域），我方全库统存 open_id ⇒ 必须转换后再进准入鉴权。为 nil 时回调若带
	// user_id 将 503 可见失败（绝不静默放行、更不得拿 user_id 冒充 open_id——假 403 红线）。
	Contact feishu.ContactClient
	// OrgSync 通讯录镜像全量同步运行器（docs/08 实施批次一）。
	// POST /internal/org/sync 手动触发入口；为 nil 时该端点返回 503 可见失败。
	OrgSync *orgsync.Runner
	// Objects 附件对象存储（主存）。为 nil 时**不缓存、直接转发**（降级，不是静默丢功能）。
	Objects objectstore.Store
	// Spec 内嵌机读规格（specload.Bundle；启动即 S1–S12 校验，见 M1）。
	// 消费方：GET /approval/meta（M3 下发表单 schema）、submit 提交校验（M4）。
	Spec *specload.Bundle
	// Chain 分档/审批链计算（FR-M9-02；chain.json 唯一权威源）。
	// submit 服务端算链（d5：调用方 nodes 一律拒绝）；preview 端点（M7）复用同一实例。
	Chain *chain.Service
	// OrgVerifier 提交实时回源端口（FR-M9-17 / M5；消费方定义，实现＝orgsync.FeishuFetcher
	// 的 contact/v3 单人接口）。为 nil 时回源跳过并落 not_assembled 标记（可见，不静默）。
	OrgVerifier OrgVerifier
}

// OrgVerifier 按 open_id 实时回源飞书通讯录单人（FR-M9-17）。
type OrgVerifier interface {
	GetUser(ctx context.Context, openID string) (*store.OrgUser, error)
}

// NewRouter 装配 Echo 路由与中间件。
func NewRouter(d Deps) *echo.Echo {
	if d.Log == nil {
		d.Log = observ.NewLogger("info", nil)
	}
	e := echo.New()
	e.HideBanner = true
	e.HidePort = true
	e.HTTPErrorHandler = d.errorHandler

	e.Use(middleware.Recover())
	e.Use(d.traceMiddleware)
	e.Use(d.logMiddleware)
	e.Use(d.rateLimitMiddleware(300)) // 简单固定窗口限流（对应 42900）

	// ---- 内部运维端点（管理凭据域，非飞书免登）----
	internal := e.Group("", d.internalAuth)
	internal.GET("/healthz", d.handleHealthz)
	internal.GET("/readyz", d.handleReadyz)
	internal.POST("/internal/sync/subscribe", d.handleSubscribe)
	internal.POST("/internal/sync/reconcile", d.handleReconcile)
	// ★ 审批对账唯一入口（R24）：对 external_instances/check 的 diff 判方向后重推。
	//   与旧 `/internal/sync/reconcile`（410 Gone）是**两个不同职能**，不重复。
	internal.POST("/internal/approval/check", d.handleApprovalCheck)
	// ★ 通讯录镜像全量同步（docs/08 实施批次一）：手动触发入口（无视新鲜度阈值）。
	internal.POST("/internal/org/sync", d.handleOrgSync)
	internal.POST("/internal/events/:id/replay", d.handleReplay)
	// ★ 开发模式注入端点：仅当 DEV_MODE=true 时注册（无凭据时的端到端验证路径）。
	if d.Env != nil && d.Env.IsDev() {
		internal.POST("/internal/dev/inject-event", d.handleInjectEvent)
	}

	// ---- 免登（公开）----
	// ★ GET /api/auth/authorize-url 挂 root `e`（公开）：未登录时前端守卫/登录页
	//   就要拿授权 URL，若进 api 组会被 requireSession 挡住（40100）死循环。
	e.GET("/api/auth/authorize-url", d.handleFeishuAuthorizeURL)
	e.GET("/auth/feishu/callback", d.handleFeishuCallback)
	e.POST("/auth/logout", d.handleLogout)

	// ---- ★ 入站回调（飞书 → 我方；独立入站面）----
	// ★★ 必须挂 **root `e`**、与 `/auth/*` 同组，**绝不进 `api` 组** —— 飞书回调请求
	//    **无会话 cookie**，若挂 `requireSession` 会 401 全失败且静默（docs/11 R14 / docs/05 §3.14）。
	//    该路径**不走** api 组、不施会话中间件；业务层凭 action_callback_token 校验（flow.HandleCallback）。
	// ★ 第三个参数＝**回调报文留痕中间件**（docs/16 §2-E / G-7）：只挂本条路由、
	//    **不全局生效**（其他路由 body 不落日志，最小化）；token 全打码 / reason 截 200 /
	//    attachments 只记条数，详见 middleware_callback.go。
	e.POST("/approval/external/callback", d.handleExternalApprovalCallback, d.callbackBodyLog)

	// ---- 业务接口（会话域，自动施加行·列过滤）----
	api := e.Group("/api", d.requireSession)
	api.GET("/me", d.handleMe)
	api.GET("/instances", d.handleListInstances)
	api.GET("/instances/:code", d.handleGetInstance)
	api.GET("/instances/:code/prefill", d.handleInstancePrefill) // B6 关联单预填（usage_category 等 source=system 带入键）
	api.GET("/instances/:code/timeline", d.handleInstanceTimeline)
	// 附件（B39）：元数据列表 + 按需下载（行级以所属实例为准）。
	api.GET("/instances/:code/attachments", d.handleInstanceAttachments)
	api.GET("/attachment/:file_id", d.handleAttachmentDownload)
	api.GET("/ledger/:table", d.handleLedgerList)
	api.GET("/ledger/:table/:id", d.handleLedgerGet)
	api.PATCH("/ledger/:table/:id", d.handleLedgerPatch)
	api.GET("/audit/logs", d.handleAuditLogs)

	// ---- 组织查询（转交/加签目标选择器数据源；★ 数据源＝t_user_role 已配置角色者，
	//      **非飞书通讯录全量**，边界说明见 handlers_org.go / docs/05-API §3.15）----
	api.GET("/org/users", d.handleOrgUsers)
	api.GET("/org/departments", d.handleOrgDepartments)

	// ---- 审批流转（架构转向 ③；T04b；docs/05-API §3.13）----
	//	★ 页面两键与入站回调**走同一状态机出口**（flow），不得两套语义（docs/11 R11）。
	//	★ 路径参数 `:biz_no` ＝**业务单号**（非 instance_id）。
	// 附件暂存（M6）：静态段先注册（Echo 静态优先，避免被 :biz_no 遮蔽）。
	api.POST("/approval/attachments", d.handleApprovalAttachmentUpload)
	api.GET("/approval/attachments/:file_id", d.handleApprovalAttachmentStagingRead)
	// 分档/链预览（M7/D5：与 submit 共算，前端不本地算分档）。
	api.POST("/approval/preview", d.handleApprovalPreview)
	// 提交：生成编号 + 建实例 + 首推飞书（薄壳，语义在 flow.Submit；docs/05-API §3.13）。
	api.POST("/approval/submit", d.handleApprovalSubmit)
	api.POST("/approval/:biz_no/approve", d.handleApprovalApprove)
	api.POST("/approval/:biz_no/backfill", d.handleApprovalBackfill) // N-062 J3 SA 侧 · 后置补录（MIMO-NEXT-BATCH-27）
	api.POST("/approval/:biz_no/reject", d.handleApprovalReject)
	// 四操作（转交 / 加签 / 回退 / 撤回）；★ 加签另带 timing ∈ {AFTER, BEFORE}（缺省 AFTER）。
	api.POST("/approval/:biz_no/transfer", d.handleApprovalTransfer)
	api.POST("/approval/:biz_no/addsign", d.handleApprovalAddSign)
	api.POST("/approval/:biz_no/rollback", d.handleApprovalRollback)
	api.POST("/approval/:biz_no/cancel", d.handleApprovalCancel)
	// 我的待办（★ 命名已定 ＝ `/tasks`，非 `/todo`）+ 定义清单（管理员）+ 单实例详情。
	//	★ 静态段（`tasks`/`defs`/`meta`）先注册、参数段（`:biz_no`）后注册（Echo 静态优先，避免遮蔽）。
	api.GET("/approval/meta", d.handleApprovalMeta)
	api.GET("/approval/tasks", d.handleApprovalTasks)
	api.GET("/approval/defs", d.handleApprovalDefs)
	api.GET("/approval/:biz_no", d.handleApprovalInstance)

	// ---- 看板（M5，只读；行级过滤在 SQL 层、列级投影在序列化层）----
	//	★ dashboard:{1..4} 权限资源按看板 id 分派，见 handlers_dashboard.go / docs/05-API.md §3.3。
	api.GET("/dashboard/:id", d.handleDashboard)
	api.GET("/dashboard/:id/export", d.handleDashboardExport)

	// ---- 备付金（M1）/ 报送（M6）----
	//	★ 两资源暂以角色口径显式鉴权（authorizeRole），不进入行·列权限矩阵，
	//	  避免出现「矩阵可配但处理器更严」的假配置；待资源枚举入库后平滑迁移。
	api.GET("/petty-cash/balance", d.handlePettyCashBalance)
	api.POST("/petty-cash/receipt", d.handlePettyCashReceipt)
	api.POST("/petty-cash/monthly-close", d.handlePettyCashMonthlyClose)
	api.POST("/submission", d.handleCreateSubmission)
	api.GET("/submission", d.handleListSubmissions)
	api.GET("/submission/:id/package", d.handleSubmissionPackage)
	api.POST("/submission/:id/receipt", d.handleRegisterSubmissionReceipt)
	api.POST("/submission/:id/group", d.handleRegisterSubmissionGroup)
	api.POST("/submission/:id/reject", d.handleRejectSubmission)

	// ---- 集团报销跟踪表（M1，FR-M1-02；同上：角色口径显式鉴权，不进权限矩阵）----
	api.POST("/reimbursement", d.handleCreateReimbursement)
	api.GET("/reimbursement", d.handleListReimbursements)
	api.PATCH("/reimbursement/:id", d.handlePatchReimbursement)

	// ---- 变更链回溯（M4，FR-M4-07；按合同号查历次变更，行级过滤在 SQL 层）----
	api.GET("/contract/:biz_no/changes", d.handleContractChanges)

	// ---- 系统管理（M5，★ 仅「系统管理员」；服务端二次校验，见 handlers_admin.go）----
	admin := api.Group("/admin")
	admin.GET("/permission-rules", d.handleAdminPermissionRulesGet)
	admin.PUT("/permission-rules", d.handleAdminPermissionRulesPut)
	admin.GET("/users", d.handleAdminUsersGet)
	admin.POST("/users", d.handleAdminUsersPost)
	admin.PATCH("/users/:open_id", d.handleAdminUsersPatch)
	// ---- 运营性常量表（T2 / R-24）：增改 + **只停用不删**（DELETE 显式拒绝）----
	admin.GET("/constants", d.handleAdminConstantsList)
	admin.POST("/constants", d.handleAdminConstantsCreate)
	admin.PUT("/constants/:id", d.handleAdminConstantsUpdate)
	admin.DELETE("/constants/:id", d.handleAdminConstantsDeleteRefused)
	// ---- 角色代理人（N-028 · 授权配置，docs/05-API §3.9）：只停用不删 + feature_enabled=true（N-072：M9 落地、解除条件已达成）----
	admin.GET("/role-agents", d.handleAdminRoleAgentsList)
	admin.POST("/role-agents", d.handleAdminRoleAgentsCreate)
	admin.PUT("/role-agents/:id", d.handleAdminRoleAgentsUpdate)
	admin.DELETE("/role-agents/:id", d.handleAdminRoleAgentsDelete)
	// ---- 三方审批定义装载（docs/16 §2-C 通道①：主通道）----
	// ★ 清单源＝`t_config_mapping(map_kind='approval_code')`（清单正本＝
	//   docs/reference/config-mapping.sample.json 的 approval_code 节，11 类）；
	//   幂等（approval_code 命中即更新）；占位符未替换 / 清单为空 ⇒ 可见错误，不静默成功。
	admin.POST("/approval/defs/sync", d.handleAdminApprovalDefsSync)

	// ---- 前端静态资源（embed 产物；缺失时 controller 降级为占位页）----
	// ★ 注意：Echo 的通配路由 "/*" 不匹配根路径 "/"，须单独注册根路由，
	//	否则访问站点根会命中 404（SPA 入口无法加载）。
	if d.WebUI != nil {
		e.GET("/", echo.WrapHandler(d.WebUI))
		e.GET("/*", echo.WrapHandler(d.WebUI))
	}
	return e
}

// ---------- 中间件 ----------

func (d Deps) traceMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
	return func(c echo.Context) error {
		id := c.Request().Header.Get("X-Request-Id")
		if strings.TrimSpace(id) == "" {
			id = newTraceID()
		}
		c.Set(ctxKeyTraceID, id)
		c.Response().Header().Set("X-API-Version", "1")
		return next(c)
	}
}

func (d Deps) logMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
	return func(c echo.Context) error {
		start := time.Now()
		err := next(c)
		d.Log.Info("http",
			"method", c.Request().Method,
			"path", c.Request().URL.Path,
			"status", c.Response().Status,
			"duration_ms", time.Since(start).Milliseconds(),
			"trace_id", traceID(c),
		)
		return err
	}
}

// internalAuth 校验内部运维端点管理凭据（X-Internal-Token）或仅回环访问。
func (d Deps) internalAuth(next echo.HandlerFunc) echo.HandlerFunc {
	return func(c echo.Context) error {
		expect := ""
		if d.Env != nil {
			expect = strings.TrimSpace(d.Env.InternalToken)
		}
		if expect != "" {
			provided := c.Request().Header.Get("X-Internal-Token")
			if subtle.ConstantTimeCompare([]byte(provided), []byte(expect)) == 1 {
				return next(c)
			}
			return fail(c, http.StatusUnauthorized, codeUnauthorized, "内部端点凭据无效")
		}
		// 未配置 token：仅允许回环地址访问（开发/本机运维）。
		if isLoopback(c.RealIP()) {
			return next(c)
		}
		return fail(c, http.StatusUnauthorized, codeUnauthorized, "内部端点未配置凭据且非本机访问")
	}
}

// requireSession 业务接口会话校验（未登录 → 40100）。
func (d Deps) requireSession(next echo.HandlerFunc) echo.HandlerFunc {
	return func(c echo.Context) error {
		ck, err := c.Cookie(access.CookieName)
		if err != nil || ck.Value == "" {
			return fail(c, http.StatusUnauthorized, codeUnauthorized, "未登录或会话失效")
		}
		sess, ok := d.Auth.ResolveSession(ck.Value)
		if !ok {
			return fail(c, http.StatusUnauthorized, codeUnauthorized, "未登录或会话失效")
		}
		c.Set(ctxKeySession, sess)
		return next(c)
	}
}

// rateLimitMiddleware 简单固定窗口限流（按 IP，1 秒窗口）。
func (d Deps) rateLimitMiddleware(perSecond int) echo.MiddlewareFunc {
	type bucket struct {
		count int
		reset time.Time
	}
	var (
		mu      sync.Mutex
		buckets = map[string]*bucket{}
	)
	return func(next echo.HandlerFunc) echo.HandlerFunc {
		return func(c echo.Context) error {
			ip := c.RealIP()
			now := time.Now()
			mu.Lock()
			b, ok := buckets[ip]
			if !ok || now.After(b.reset) {
				b = &bucket{count: 0, reset: now.Add(time.Second)}
				buckets[ip] = b
			}
			b.count++
			exceeded := b.count > perSecond
			mu.Unlock()
			if exceeded {
				return fail(c, http.StatusTooManyRequests, codeRateLimited, "请求过于频繁")
			}
			return next(c)
		}
	}
}

// errorHandler 将错误统一包裹为 Envelope。
func (d Deps) errorHandler(err error, c echo.Context) {
	if c.Response().Committed {
		return
	}
	status := http.StatusInternalServerError
	code := codeInternal
	msg := "服务器内部错误"

	var he *echo.HTTPError
	if errors.As(err, &he) {
		status = he.Code
		switch he.Code {
		case http.StatusNotFound:
			code, msg = codeNotFound, "资源不存在"
		case http.StatusUnauthorized:
			code, msg = codeUnauthorized, "未登录或会话失效"
		case http.StatusForbidden:
			code, msg = codeForbidden, "无权限"
		case http.StatusTooManyRequests:
			code, msg = codeRateLimited, "请求过于频繁"
		default:
			code = he.Code * 100
			if s, ok := he.Message.(string); ok {
				msg = s
			}
		}
	}
	d.Log.Error("请求处理失败", "path", c.Request().URL.Path, "error", err.Error())
	_ = c.JSON(status, Envelope{Code: code, Data: nil, Message: msg, TraceID: traceID(c)})
}

func isLoopback(ip string) bool {
	host := strings.TrimSpace(ip)
	if h, _, err := net.SplitHostPort(host); err == nil {
		host = h
	}
	if host == "127.0.0.1" || host == "::1" || host == "localhost" {
		return true
	}
	parsed := net.ParseIP(host)
	return parsed != nil && parsed.IsLoopback()
}