# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "采购与费用审批平台（自建侧）· 接口契约（机读形态）",
    "version": "1.0.0",
    "description": "★ 本文件是 `docs/05-API.md`（**人读正本**）的**机读形态**，由 `scripts/gen_openapi.py` **机械生成** —— **禁止手工编辑**；改接口请改正本后 `python scripts/gen_openapi.py` 重生。\n★ 本文件**不新增、不放松任何判据**：凡是正本没写的（逐端点请求/响应字段类型、枚举值域、必填性），本文件一律**不填**（见 `x-known-gaps`），由正本的表格文字承载。\n★ 一致性由两处**独立**把关：门禁「会报」项 `C9`（`scripts/audit_silent.py`：本文件 ↔ 正本**双向**路由差集）与常驻探针 `scripts/_probe_n051.py`（重现一致性 ＋ 路径参数完备性 ＋ operationId 唯一性 ＋ 正反例）。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署（前端 `internal/webui/dist` 内嵌于同一二进制）"
    }
  ],
...
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
    "/api/auth/authorize-url": {
      "get": {
        "operationId": "get_api_auth_authorize_url",
        "tags": [
          "3.1 鉴权"
        ],
        "summary": "下发飞书免登**授权页 URL**：生成随机 state（crypto/rand 16 字节，存 HttpOnly Cookie jx_oauth_state，10 分钟）→ 按官方拼装 https://accounts.feishu.cn/open-apis/authen/v1/authorize?client_id=…&response_type=code&redirect_uri=<URL编码>&state=…",
        "security": [],
        "x-auth-domain": "公开（不进 `requireSession` 组，§3.1）",
        "x-error-codes": [
          40000,
          50000
        ],
        "x-related-fr": [
          "FR-M0-10"
        ],
        "x-doc-ref": "docs/05-API.md#L123-L134",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "错误码 40000（HTTP 400；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "错误码 50000（HTTP 500；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/feishu/callback": {
      "get": {
        "operationId": "get_auth_feishu_callback",
        "tags": [
          "3.1 鉴权"
        ],
        "summary": "飞书免登回调",
        "security": [],
        "x-auth-domain": "公开（不进 `requireSession` 组，§3.1）",
        "x-error-codes": [
          40000,
          40101
        ],
        "x-related-fr": [
          "FR-M0-10"
        ],
        "x-doc-ref": "docs/05-API.md#L135-L145",
        "responses": {
          "302": {
            "description": "登录成功后 302 回跳原目标页（`jx_oauth_redirect`；取不到/非法 ⇒ 回落 `/`）；失败/拒绝授权同样 302 至登录错误页（见 §3.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "错误码 40000（HTTP 400；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "错误码 40101（HTTP 401；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "post_auth_logout",
        "tags": [
          "3.1 鉴权"
        ],
        "summary": "注销会话",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "x-error-codes": [
          40100
        ],
        "x-related-fr": [
          "FR-M0-10"
        ],
        "x-doc-ref": "docs/05-API.md#L146-L156",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "错误码 40100（HTTP 401；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "operationId": "get_api_me",
        "tags": [
          "3.1 鉴权"
        ],
        "summary": "返回当前会话身份、角色、可见范围摘要（供前端展示降级）",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "x-error-codes": [
          40100,
          40101
        ],
        "x-related-fr": [
          "FR-M0-10",
          "FR-M5-04"
        ],
        "x-doc-ref": "docs/05-API.md#L157-L167",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "错误码 40100 / 40101（HTTP 401；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/{table}": {
      "get": {
        "operationId": "get_api_ledger_table",
        "tags": [
          "3.2 台账（M4）"
        ],
        "summary": "台账列表查询",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-error-codes": [
          40300,
          40301,
          40000
        ],
        "x-related-fr": [
          "FR-M4-01",
          "FR-M4-04",
          "FR-M4-05",
          "FR-M5-02",
          "FR-M5-03",
          "FR-M7-01"
        ],
        "x-doc-ref": "docs/05-API.md#L170-L183",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "错误码 40000（HTTP 400；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "错误码 40300 / 40301（HTTP 403；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/{table}/{id}": {
      "get": {
        "operationId": "get_api_ledger_table_id",
        "tags": [
          "3.2 台账（M4）"
        ],
        "summary": "单条台账记录查询",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-error-codes": [
          40301,
          40400
        ],
        "x-related-fr": [
          "FR-M5-02",
          "FR-M5-03",
          "FR-M7-01"
        ],
        "x-doc-ref": "docs/05-API.md#L184-L194",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "错误码 40301（HTTP 403；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "错误码 40400（HTTP 404；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch_api_ledger_table_id",
        "tags": [
          "3.2 台账（M4）"
        ],
        "summary": "**仅运营表可写字段**的登记/回填（付款凭据号、抽查状态、经办状态、集团侧人工字段等）",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-error-codes": [
          40300,
          40901,
          40000
        ],
        "x-related-fr": [
          "FR-M1-03",
          "FR-M4-03",
          "FR-M4-09",
          "FR-M6-07"
        ],
        "x-doc-ref": "docs/05-API.md#L195-L210",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "错误码 40000（HTTP 400；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "错误码 40300（HTTP 403；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "错误码 40901（HTTP 409；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/{id}": {
      "get": {
        "operationId": "get_api_dashboard_id",
        "tags": [
          "3.3 看板（M5）"
        ],
        "summary": "看板指标数据",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-error-codes": [
          40300,
          50300
        ],
        "x-related-fr": [
          "FR-M5-01",
          "FR-M5-05",
          "FR-M5-06",
          "FR-M5-07",
          "FR-M5-08"
        ],
        "x-doc-ref": "docs/05-API.md#L213-L224",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "错误码 40300（HTTP 403；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "错误码 50300（HTTP 503；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/{id}/export": {
      "get": {
        "operationId": "get_api_dashboard_id_export",
        "tags": [
          "3.3 看板（M5）"
        ],
        "summary": "看板数据导出",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-error-codes": [
          40300,
          42900
        ],
        "x-related-fr": [
          "FR-M7-03",
          "FR-M5-06"
        ],
        "x-doc-ref": "docs/05-API.md#L225-L236",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "错误码 40300（HTTP 403；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "错误码 42900（HTTP 429；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances": {
      "get": {
        "operationId": "get_api_instances",
        "tags": [
          "3.4 实例（M2）"
        ],
        "summary": "实例列表查询",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "x-error-codes": [
          40300,
          40301
        ],
        "x-related-fr": [
          "FR-M2-01",
          "FR-M2-03",
          "FR-M2-05",
          "FR-M5-02",
          "FR-M5-03"
        ],
        "x-doc-ref": "docs/05-API.md#L239-L249",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "错误码 40300 / 40301（HTTP 403；含义见 docs/05-API.md §2.1 错误码表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/approval/meta": {
      "get": {
        "operationId": "get_api_approval_meta",
        "tags": [
          "3.13 审批流转（我方页面 · **转向 ③ 新增**）"
        ],
        "summary": "下发表单 schema（spec/forms/*.json 内嵌）→ 前端 meta 驱动渲染",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-auth-domain": "飞书免登会话（§1.2；`open_id` 未映射 ⇒ 40101）",
        "x-related-fr": [
          "FR-M9-03"
        ],
        "x-doc-ref": "docs/05-API.md#L678-L686",
        "responses": {
          "200": {
            "description": "成功（`code=0`）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "未列举的错误码（统一错误包裹 `code != 0`；完整码表见 docs/05-API.md §2.1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jx_session",
        "description": "飞书免登会话 Cookie（HttpOnly / Secure / SameSite=Lax，§1.2）。★ 角色**不缓存在会话里**，每请求实时解析（`t_user_role`），停用/改角色即时生效。"
      },
      "internalToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Token",
        "description": "内部运维端点管理凭据 `JX_INTERNAL_TOKEN`（§3.8）。★ **不得**再依赖「仅绑定回环/内网」做鉴权（`04a §17.6`）——回环绑定只作纵深防御、**不替代 token**。"
      }
    },
...
  },
...
}